feat(initial): Initial version of DoH blocklist creator
This isn't fully functional yet, but it's to the point where I want to
start getting some of the code in version control. Right now this
program can fetch a set of lists from either local files or through
HTTP(S) that contain lists of hosts. It then dedupes those lists and
finally looks up the IPv4 and IPv6 addresses for the hosts and stores
them in two separate lists.

The next step is to make it so the program can log into the UDM Pro and
update those IP Groups accordingly. From there, it's pretty trivial to
create a firewall rule in the UDM Pro to block most of the machines on
your network from accessing those hosts.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>