feat(unifi): Properly updates groups on UDM-Pro
This now properly updates the groups on my Unifi Dream Machine Pro. It's
still not quite where I want it because the groups need to already
exist. It also is hardcoded to use the default site. There's a bit of
scaffolding that should allow me to work around that in the future.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>