fix(CSRF): fix for PUT requests
At some point the API changed so PUT requests, such as updating a
firewall group, required a CSRF token. This makes it so the CSRF token
is not blown away if not returned and therefore persists for multiple
calls after a login.

Closes pridkett/unifi-doh-blocker#1

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>